Caesar.Cipher (caesar v0.1.0)

Caesar Cipher module, which maps characters to new characters. This program is a demostration of my understanding of Elixir programming language. The code pattern is influenced by my learning source.

Functions

calculate_mapping(base_letter, char, shift)

encrypt(msg, shift)

Encrypt the message by applying a mapping that shifts the alphabet by the shift value.

Examples

iex> Caesar.Cipher.encrypt("hello world!", 1)
"gdkkn vnqkc!"